income tax nics alignment (#95)
* income tax nics alignment

* PR review changes

+++ app/v1/models/response/getIncomeTaxAndNics/NicBand.scala
@@ -20,9 +20,9 @@ import play.api.libs.json.{Json, OFormat}
 
 case class NicBand(name: String,
                    rate: BigDecimal,
-                   threshold: Option[BigDecimal],
-                   apportionedThreshold: Option[BigDecimal],
-                   income: BigDecimal,
+                   threshold: Option[BigInt],
+                   apportionedThreshold: Option[BigInt],
+                   income: BigInt,
                    amount: BigDecimal)
 
 object NicBand {
